+++ air/src/lib.rs
@@ -29,6 +29,7 @@ pub mod ace;
 pub mod config;
 mod constraints;
 pub mod lookup;
+pub mod memory;
 mod proof_order;
 pub mod trace;
 
@@ -99,6 +100,7 @@ mod export {
             MultiAir, PermutationAirBuilder, ProverStatement, Statement,
         },
         debug,
+        pcs::PcsParams,
     };
 }